refactor(ssm_document): address PR review comments
- Replace VerifyEnvConfigContent SSM round-trip with direct os.ReadFile
  of the platform-appropriate env-config.json path (test runs on instance)
- Add overwrite test case asserting updated value on disk after set-env
  of an existing key
- Add allowedPattern-rejection test case asserting SendCommand returns
  InvalidParameters for a value containing '$' (rejected by pattern)
- De-duplicate Validate() bodies into shared validate.go; platform files
  now only carry go:embed vars, envConfigPath const, and platformSetup()
  behind build tags
