+++ test/ssm_document/ssm_document_unix.go
@@ -11,13 +11,6 @@ import (
 	"log"
 	"os/exec"
 	"strings"
-	"time"
-
-	"github.com/aws/aws-sdk-go-v2/service/ssm/types"
-	"github.com/google/uuid"
-
-	"github.com/aws/amazon-cloudwatch-agent-test/environment"
-	"github.com/aws/amazon-cloudwatch-agent-test/util/awsservice"
 )
 
 var (
@@ -29,199 +22,25 @@ var (
 	agentConfig2 string
 )
 
-// set-env on-disk verification: read the agent's env-config.json through the stock
-// run-command document for this platform.
-const (
-	runCommandDocumentName = "AWS-RunShellScript"
-	envConfigReadCommand   = "cat /opt/aws/amazon-cloudwatch-agent/etc/env-config.json"
-)
+// envConfigPath is the on-disk path to the agent's env-config.json on Linux/macOS.
+const envConfigPath = "/opt/aws/amazon-cloudwatch-agent/etc/env-config.json"
 
-func Validate() error {
-	log.Println("Starting SSM Document validation tests")
-
-	// Verify shell compatibility and log shell type
-	if err := VerifyShellCompatibility(); err != nil {
+// platformSetup performs platform-specific initialization (shell compatibility check on unix).
+func platformSetup() {
+	if err := verifyShellCompatibility(); err != nil {
 		log.Printf("Warning: Shell compatibility verification failed: %v", err)
 	}
-
-	// Generate unique ID to guarantee uniqueness
-	uniqueID := uuid.New().String()[:8]
-	documentName := testManageAgentDocument + uniqueID
-	agentConfig1Name := agentConfigFile1 + "-" + uniqueID
-	agentConfig2Name := agentConfigFile2 + "-" + uniqueID
-	metadata := environment.GetEnvironmentMetaData()
-	instanceIds := []string{metadata.InstanceId}
-
-	// Wait for SSM agent to be ready before running tests
-	log.Printf("Waiting for instance %s to be SSM-ready...", metadata.InstanceId)
-	if err := awsservice.WaitForSSMReady(instanceIds, 5*time.Minute); err != nil {
-		return fmt.Errorf("instance not SSM-ready: %v", err)
-	}
-	log.Println("Instance is SSM-ready")
-
-	log.Printf("Creating SSM document: %s", documentName)
-	err := awsservice.CreateSSMDocument(documentName, manageAgentDoc, types.DocumentTypeCommand)
-	if err != nil {
-		return err
-	}
-
-	// Ensure SSM document is cleaned up regardless of test outcome
-	defer func() {
-		log.Printf("Cleaning up SSM document: %s", documentName)
-		if deleteErr := awsservice.DeleteSSMDocument(documentName); deleteErr != nil {
-			log.Printf("Warning: Failed to delete SSM document %s: %v", documentName, deleteErr)
-		} else {
-			log.Printf("Successfully deleted SSM document: %s", documentName)
-		}
-	}()
-
-	// Test start action
-	startTest := testCase{
-		parameters:           map[string][]string{paramAction: {actionStart}},
-		actionName:           actionStart,
-		expectedAgentStatus:  agentStatusRunning,
-		expectedConfigStatus: configStatusConfigured,
-	}
-	if err := RunAndVerifySSMAction(documentName, instanceIds, startTest); err != nil {
-		return err
-	}
-
-	// Test stop action
-	stopTest := testCase{
-		parameters:           map[string][]string{paramAction: {actionStop}},
-		actionName:           actionStop,
-		expectedAgentStatus:  agentStatusStopped,
-		expectedConfigStatus: configStatusConfigured,
-	}
-	if err := RunAndVerifySSMAction(documentName, instanceIds, stopTest); err != nil {
-		return err
-	}
-
-	// Test configure (remove) action
-	removeTest := testCase{
-		parameters: map[string][]string{
-			paramAction:                      {actionConfigureRemove},
-			paramOptionalConfigurationSource: {configSourceAll},
-			paramOptionalRestart:             {restartNo},
-		},
-		actionName:           actionConfigureRemove,
-		expectedAgentStatus:  agentStatusStopped,
-		expectedConfigStatus: configStatusNotConfigured,
-	}
-	if err := RunAndVerifySSMAction(documentName, instanceIds, removeTest); err != nil {
-		return err
-	}
-
-	// Test configure action
-	log.Printf("Putting SSM parameter: %s", agentConfig1Name)
-	if err := awsservice.PutStringParameter(agentConfig1Name, agentConfig1); err != nil {
-		return err
-	}
-	defer cleanupSSMParameter(agentConfig1Name)
-
-	configureTest := testCase{
-		parameters: map[string][]string{
-			paramAction:                        {actionConfigure},
-			paramOptionalConfigurationSource:   {configSourceSSM},
-			paramOptionalConfigurationLocation: {agentConfig1Name},
-		},
-		actionName:           actionConfigure,
-		expectedAgentStatus:  agentStatusRunning,
-		expectedConfigStatus: configStatusConfigured,
-	}
-	if err := RunAndVerifySSMAction(documentName, instanceIds, configureTest); err != nil {
-		return err
-	}
-
-	// Test configure (append) action
-	log.Printf("Putting SSM parameter: %s", agentConfig2Name)
-	if err := awsservice.PutStringParameter(agentConfig2Name, agentConfig2); err != nil {
-		return err
-	}
-	defer cleanupSSMParameter(agentConfig2Name)
-
-	appendTest := testCase{
-		parameters: map[string][]string{
-			paramAction:                        {actionConfigureAppend},
-			paramOptionalConfigurationSource:   {configSourceSSM},
-			paramOptionalConfigurationLocation: {agentConfig2Name},
-		},
-		actionName:           actionConfigureAppend,
-		expectedAgentStatus:  agentStatusRunning,
-		expectedConfigStatus: configStatusConfigured,
-	}
-	if err := RunAndVerifySSMAction(documentName, instanceIds, appendTest); err != nil {
-		return err
-	}
-
-	// Test set-env action (happy path): custom (non translator-managed) key with a value
-	// containing spaces. Verifies the ctl's "Set <KEY>" output, that the pair is persisted
-	// to env-config.json on disk, and that agent status/configstatus are unchanged.
-	setEnvTest := testCase{
-		parameters: map[string][]string{
-			paramAction:                      {actionSetEnv},
-			paramOptionalEnvironmentVariable: {setEnvKey1 + "=" + setEnvValue1},
-		},
-		actionName:           actionSetEnv,
-		expectedAgentStatus:  agentStatusRunning,
-		expectedConfigStatus: configStatusConfigured,
-	}
-	if err := RunAndVerifySSMActionWithOutput(documentName, instanceIds, setEnvTest, setEnvOutputPrefix+setEnvKey1); err != nil {
-		return err
-	}
-	if err := VerifyEnvConfigContent(runCommandDocumentName, envConfigReadCommand, metadata.InstanceId, map[string]string{
-		setEnvKey1: setEnvValue1,
-	}); err != nil {
-		return err
-	}
-
-	// Test set-env action (merge): a second key is persisted alongside the first.
-	setEnvMergeTest := testCase{
-		parameters: map[string][]string{
-			paramAction:                      {actionSetEnv},
-			paramOptionalEnvironmentVariable: {setEnvKey2 + "=" + setEnvValue2},
-		},
-		actionName:           actionSetEnvMerge,
-		expectedAgentStatus:  agentStatusRunning,
-		expectedConfigStatus: configStatusConfigured,
-	}
-	if err := RunAndVerifySSMActionWithOutput(documentName, instanceIds, setEnvMergeTest, setEnvOutputPrefix+setEnvKey2); err != nil {
-		return err
-	}
-	if err := VerifyEnvConfigContent(runCommandDocumentName, envConfigReadCommand, metadata.InstanceId, map[string]string{
-		setEnvKey1: setEnvValue1,
-		setEnvKey2: setEnvValue2,
-	}); err != nil {
-		return err
-	}
-
-	// Test set-env action (error path): empty optionalEnvironmentVariable must fail
-	// with the document-level error message.
-	setEnvEmptyTest := testCase{
-		parameters: map[string][]string{
-			paramAction:                      {actionSetEnv},
-			paramOptionalEnvironmentVariable: {""},
-		},
-		actionName: actionSetEnvEmpty,
-	}
-	if err := RunAndVerifySSMActionFailure(documentName, instanceIds, setEnvEmptyTest, setEnvEmptyErrorMessage); err != nil {
-		return err
-	}
-
-	log.Println("All SSM Document validation tests completed successfully")
-	return nil
 }
 
-// ShellInfo contains information about the detected shell
+// shellInfo contains information about the detected shell.
 type shellInfo struct {
 	shellPath string
 	shellType string
 	isPOSIX   bool
 }
 
-// getShellType returns the shell type for /bin/sh
+// getShellType returns the shell type for /bin/sh.
 func getShellType() (*shellInfo, error) {
-	// Use readlink to resolve the /bin/sh symlink
 	cmd := exec.Command("readlink", "-f", "/bin/sh")
 	output, err := cmd.Output()
 	if err != nil {
@@ -232,7 +51,6 @@ func getShellType() (*shellInfo, error) {
 	shellType := "unknown"
 	isPOSIX := false
 
-	// Determine shell type based on the resolved path
 	if strings.Contains(shellPath, "dash") {
 		shellType = "dash"
 		isPOSIX = true
@@ -240,7 +58,6 @@ func getShellType() (*shellInfo, error) {
 		shellType = "bash"
 		isPOSIX = true
 	} else if strings.Contains(shellPath, "sh") {
-		// Generic sh, assume POSIX-compliant
 		shellType = "sh"
 		isPOSIX = true
 	}
@@ -252,19 +69,19 @@ func getShellType() (*shellInfo, error) {
 	}, nil
 }
 
-// VerifyShellCompatibility checks if the system shell is compatible and logs the information
-func VerifyShellCompatibility() error {
-	shellInfo, err := getShellType()
+// verifyShellCompatibility checks if the system shell is compatible and logs the information.
+func verifyShellCompatibility() error {
+	info, err := getShellType()
 	if err != nil {
 		return fmt.Errorf("shell compatibility check failed: %w", err)
 	}
 
 	log.Printf("Shell compatibility check:")
-	log.Printf("  /bin/sh resolves to: %s", shellInfo.shellPath)
-	log.Printf("  Detected shell type: %s", shellInfo.shellType)
-	log.Printf("  POSIX-compliant: %v", shellInfo.isPOSIX)
+	log.Printf("  /bin/sh resolves to: %s", info.shellPath)
+	log.Printf("  Detected shell type: %s", info.shellType)
+	log.Printf("  POSIX-compliant: %v", info.isPOSIX)
 
-	if !shellInfo.isPOSIX {
+	if !info.isPOSIX {
 		log.Printf("WARNING: Shell may not be POSIX-compliant")
 	}
